Refactored Get Methods into sub-modules that handle the requests based on versioned modules.
